core_auth/pyproject.toml
[tool.poetry]
name = "diag-sys-auth"
version = "0.1.0"
description = ""
authors = ["Dennis <dennis.arcelopez@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "diag_sys_auth"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-split-settings = "^1.2.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pre-commit = "^3.3.2"

[tool.isort]
multi_line_outpit = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
